Move template methods out of `ParserHelper.h`
These templates methods are only used in `Configuration`. Further, we probably won't be pursuing more general use of these template methods, so it makes sense to limit their visibility. One issue with template methods is they are implemented in header files, and require those header files to include anything needed for the function implementation. The added transitive includes can have a negative impact on compile times.
